Refactor and Cleanup
