--- lib/cash.dart
@@ -1,5 +1,0 @@
-class Cash {
-  int limit;
-  Map<String, int> cashPerDay;
-  Cash({required this.cashPerDay, this.limit = 100});
-}
