+++ lib/entities/cash.dart
@@ -1,0 +1,4 @@
+class Cash {
+  Map<String, int> cashPerDay;
+  Cash({required this.cashPerDay});
+}
